Add mail-client/mblaze.
This is similar to the existing ebuild in the qsx overlay, but adds a
USE flag to include the additional tools from the contrib/ directory
(the existing ebuild only incorporates msearch). Along with this,
instead of being included by default, zsh completion is enabled only
with the zsh-completion USE flag.